Fix script and resync
